feat(server): Add autostart support in config file
- Added autostart to configuration file
- Only jobs with autostart set to true are automatically started

+++ internal/manager/manager.go
@@ -10,6 +10,7 @@ import (
 
 type JobManager struct {
 	Jobs []*job.Job
+	Config *config.Config
 }
 
 func Init(conf config.Config) *JobManager {
@@ -21,17 +22,34 @@ func Init(conf config.Config) *JobManager {
 	}
 
 	manager.Jobs = jobs
+
+	manager.Config = &conf
+
 	return &manager
 }
 
 func (ins *JobManager) Start() {
 	for _, j := range ins.Jobs {
+		autostart := ins.Config.Programs[j.Name].Autostart
+
+		if autostart == false {
+			continue
+		}
+
 		j.StartJob()
 	}
 }
 
 func (ins *JobManager) Finish() {
 	for _, j := range ins.Jobs {
+		// HACK: This is a temporary fix
+		// - Jobs should have a state (Started, Stopped, Running, ...etc)
+		// - Status must be checked before trying to finish a job
+		autostart := ins.Config.Programs[j.Name].Autostart
+		if !autostart {
+			continue
+		}
+
 		fmt.Fprintf(os.Stdout, "Exiting [%s]\n", j.Name)
 		err := j.Command.Process.Kill()
 		if err != nil {
